🎨 Renames states folder
