+++ esp32/include/motion_states/rest_state.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <gait/state.h>
+#include <motion_states/state.h>
 
 class RestState : public MotionState {
   protected:
