+++ esp32/include/motion_states/stand_state.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <gait/state.h>
+#include <motion_states/state.h>
 
 class StandState : public MotionState {
   protected:
